Test: Move test_suite_err_t to new header
The error codes are used by the NS test applications, but they are not
exported currently. Moving the definition of the error code test_suite_err_t
from `test/framework/test_framework.h` to
`test/framework/test_framework_error_codes.h` allows exporting these
error codes.
