+++ test/framework/test_framework.h
@@ -14,6 +14,7 @@
 
 #include "tfm_log_raw.h"
 #include "test_framework_helpers.h"
+#include "test_framework_error_codes.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -64,17 +65,6 @@ struct test_suite_t {
     uint32_t list_size;            /*!< List size */
     const char *name;              /*!< Test suite name */
     enum test_status_t val;        /*!< Test suite result \ref test_result_t */
-};
-
-enum test_suite_err_t {
-    TEST_SUITE_ERR_NO_ERROR = 0,           /*!< No error */
-    TEST_SUITE_ERR_INVALID_DATA = 1,       /*!< Invalid test suite if any of the
-                                            *   pointers is NULL
-                                            */
-    TEST_SUITE_ERR_INVALID_TEST_DATA = 2,  /*!< Invalid test if any of the
-                                            *  pointers is NULL
-                                            */
-    TEST_SUITE_ERR_TEST_FAILED  = 3,       /*!< Last executed test has failed */
 };
 
 /**
